Refactor assignment tests into seperate folder
